adjusted for experiment

+++ forte/topy_server.py
@@ -16,7 +16,6 @@ sys.path.append('./scripts')
 import subprocess
 import json
 import math
-import datetime
 
 from nodenums import node_nums_3d
 from nodenums import elm_num_3d
@@ -28,15 +27,14 @@ EPSILON = 1e-9
 ANALYSIS = 0
 OPTIMIZATION = 1
 
-tpd_template = '{"PROB_TYPE":"comp", "PROB_NAME":"NONAME", "ETA": "0.4", "DOF_PN": "3", "VOL_FRAC": "0.3", "FILT_RAD": "1.5", "ELEM_K": "H8", "NUM_ELEM_X":"10", "NUM_ELEM_Y":"10", "NUM_ELEM_Z":"10", "NUM_ITER":"50", "FXTR_NODE_X":"", "FXTR_NODE_Y":"", "FXTR_NODE_Z":"", "LOAD_NODE_X":"", "LOAD_VALU_X":"", "LOAD_NODE_Y":"", "LOAD_VALU_Y":"", "LOAD_NODE_Z":"", "LOAD_VALU_Z":"", "P_FAC":"1", "P_HOLD":"15", "P_INCR":"0.2", "P_CON":"1", "P_MAX":"3", "Q_FAC":"1", "Q_HOLD":"15", "Q_INCR":"0.05", "Q_CON":"1", "Q_MAX":"5"}';
-
-test_data_03 = {"forte":
-['{"boundaries":[[[-31.75,-74.81,0.5],[-29.6,-74.45,0.5],[-26.73,-74.45,0.5],[-23.14,-74.45,0.5],[-19.91,-74.45,0.5],[-17.04,-74.45,0.5],[-13.81,-74.45,0.5],[-10.58,-74.81,0.5],[-8.07,-75.16,0.5],[-4.48,-75.16,0.5],[-2.69,-75.52,0.5],[0.18,-75.88,0.5],[1.61,-76.24,0.5],[4.13,-76.6,0.5],[6.46,-76.96,0.5],[8.43,-76.96,0.5],[10.23,-76.96,0.5],[12.38,-76.96,0.5],[14.17,-76.96,0.5],[16.5,-77.32,0.5],[18.66,-77.68,0.5],[21.17,-77.68,0.5],[24.55,-77.68,0.5],[24.94,-77.68,0.5]]],"design":[{"node1":[-11.83,-22.79,4.33],"node2":[-54.35,-2.69,0.5],"points":[[-11.83,-22.79,4.33],[-15.97,-23.14,0.5],[-20.27,-19.91,0.5],[-24.58,-17.4,0.5],[-29.6,-14.89,0.5],[-34.98,-12.02,0.5],[-41.44,-9.51,0.5],[-46.46,-7.35,0.5],[-50.41,-5.56,0.5],[-53.95,-2.96,0.5],[-54.35,-2.69,0.5],[-54.35,-2.69,0.5]],"thickness":[5,5,5,5,5,5,5,5,5,5,5,6.050000000000001]},{"node1":[-0.9,-69.78,0.5],"node2":[1.95,-50.76,4.03],"points":[[-0.9,-69.78,0.5],[-0.9,-69.78,0.5],[-0.9,-67.81,0.5],[-0.9,-65.12,0.5],[-0.9,-61.53,0.5],[-0.9,-57.94,0.5],[-0.9,-54,0.5],[1.95,-50.76,4.03]],"thickness":[6.050000000000001,5,5,5,5,5,5,5]},{"node1":[1.95,-50.76,4.03],"node2":[-11.83,-22.79,4.33],"points":[[1.95,-50.76,4.03],[-3.05,-45.39,0.5],[-4.48,-41.44,0.5],[-6.28,-37.13,0.5],[-8.07,-33.19,0.5],[-10.23,-29.6,0.5],[-11.83,-22.79,4.33]],"thickness":[5,5,5,5,5,5,5]},{"node1":[1.95,-50.76,4.03],"node2":[66.78,-30.26,1],"points":[[1.95,-50.76,4.03],[1.95,-50.76,4.03],[1.97,-51.02,1],[2.69,-49.59,1],[4.83,-47.44,1],[7.34,-44.94,1],[9.13,-42.79,1],[11.64,-40.64,1],[14.5,-38.49,1],[16.65,-37.06,1],[20.23,-35.98,1],[22.38,-35.27,1],[24.17,-34.55,1],[26.32,-34.19,1],[28.82,-33.48,1],[33.12,-33.12,1],[36.7,-32.76,1],[42.43,-32.05,1],[47.8,-31.33,1],[52.1,-30.97,1],[56.04,-30.97,1],[58.18,-30.97,1],[60.69,-30.97,1],[62.84,-30.97,1],[66.33,-30.44,1],[66.78,-30.26,1],[66.78,-30.26,1]],"thickness":[5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,6.050000000000001]},{"node1":[-11.83,-22.79,4.33],"node2":[22.75,14.63,4.76],"points":[[-11.83,-22.79,4.33],[-11.83,-22.79,4.33],[-11.64,-22.74,1],[-10.2,-20.23,1],[-7.34,-16.29,1],[-4.12,-12.35,1],[0.18,-6.98,1],[4.83,-1.25,1],[8.06,2.69,1],[10.2,5.19,1],[12.71,7.34,1],[15.22,9.31,1],[17.9,10.56,1],[22.75,14.63,4.76]],"thickness":[5,5,5,5,5,5,5,5,5,5,5,5,5,5]},{"node1":[22.75,14.63,4.76],"node2":[82.17,14.86,1],"points":[[22.75,14.63,4.76],[27.03,11.64,1],[33.12,11.99,1],[41.71,12.71,1],[49.95,13.79,1],[60.33,14.5,1],[67.85,14.86,1],[74.3,14.86,1],[78.95,14.86,1],[82.17,14.86,1],[82.17,14.86,1]],"thickness":[5,5,5,5,5,5,5,5,5,5,6.050000000000001]},{"node1":[5.92,40.72,0.5],"node2":[-48.26,49.33,0.5],"points":[[5.92,40.72,0.5],[5.92,40.72,0.5],[-9.54,44.23,-3.02],[-23.04,46.17,-1.85],[-36.9,47.75,-0.67],[-48.26,49.33,0.5],[-48.26,49.33,0.5]],"thickness":[5,5,5,5,5,5,6.050000000000001]},{"node1":[22.75,14.63,4.76],"node2":[5.88,40.48,2],"points":[[22.75,14.63,4.76],[22.75,14.63,4.76],[22.69,15.63,1.5],[22.69,17.33,1.5],[22.69,19.12,1.5],[21.98,21.62,1.5],[21.98,24.12,1.5],[21.62,25.91,1.5],[20.73,28.59,1.5],[19.48,31.98,1.5],[17.51,34.3,1.5],[14.83,36.27,1.5],[11.97,37.34,1.5],[9.11,38.06,1.5],[5.88,40.48,2]],"thickness":[5,5,5,5,5,5,5,5,5,5,5,5,5,5,5]}],"loads":[{"points":[[-33.14,44.71,6.12],[-25.26,46.1,-2.43],[-9.46,43.14,1],[-18.9,44.76,2],[-15.33,44.22,2],[-11.59,43.69,2],[-8.74,43.33,2],[-9.47,42.52,1.5],[-6.6,43.15,2],[-4.46,42.79,2],[-2.67,42.44,2],[0,41.72,2],[2.5,41.19,2],[4.1,40.34,1.75],[3.74,41.01,2]],"vectors":[[0,-0.01,0],[0,-0.01,0],[0,-0.01,0],[0,-0.01,0],[0,-0.01,0],[0,-0.01,0],[0,-0.01,0],[0,-0.01,0],[0,-0.01,0],[0,-0.01,0],[0,-0.01,0],[0,-0.01,0],[-0.01,-0.01,0],[-0.01,-0.01,0],[-0.01,-0.01,0]]},{"points":[[35.99,13.52,5.81],[37.42,12.35,1],[45.83,13.25,1],[55.14,14.14,1],[64.09,14.68,1],[71.07,14.86,1]],"vectors":[[0.01,-0.04,0],[0.01,-0.04,0],[0,-0.04,0],[0,-0.04,0],[0,-0.04,0],[0,-0.04,0]]},{"points":[[-48.71,-5.45,5.42],[-48.44,-6.46,0.5],[-43.95,-8.43,0.5],[-38.21,-10.76,0.5],[-32.29,-13.45,0.5],[-27.09,-16.15,0.5],[-22.42,-18.66,0.5]],"vectors":[[-0.01,-0.04,0],[-0.01,-0.04,0],[-0.02,-0.04,0],[-0.02,-0.04,0],[-0.02,-0.04,0],[-0.02,-0.03,0],[-0.02,-0.03,0]]},{"points":[[23,-36.34,5.83],[23.27,-34.91,1],[25.24,-34.37,1],[27.57,-33.84,1],[30.97,-33.3,1],[34.91,-32.94,1],[39.57,-32.4,1],[45.12,-31.69,1],[49.95,-31.15,1],[54.07,-30.97,1],[57.11,-30.97,1]],"vectors":[[0.01,-0.02,0],[0.01,-0.02,0],[0.01,-0.02,0],[0.01,-0.02,0],[0.01,-0.02,0],[0.01,-0.02,0],[0,-0.02,0],[0,-0.02,0],[0,-0.02,0],[0,-0.02,0],[0,-0.02,0]]}],"clearances":[[[21.09,72.87,9.03],[21.14,73.57,4.08],[15.29,38.72,4.11],[15.35,39.43,-0.84],[-40.17,83.86,4.81],[-40.23,83.16,9.76],[-45.96,49.72,-0.11],[-46.02,49.02,4.84]],[[79.61,55.38,1.67],[79.62,55.15,-3.33],[82.53,16.3,3.45],[82.54,16.08,-1.54],[24.83,51.06,-3.24],[24.82,51.28,1.76],[27.75,11.98,-1.45],[27.74,12.21,3.54]],[[4.71,21.33,1.63],[4.67,21.16,-3.37],[-14.55,-21.34,3.18],[-14.58,-21.51,-1.81],[-42.61,42.49,-3.74],[-42.58,42.65,1.26],[-61.87,-0.18,-2.18],[-61.84,-0.02,2.81]],[[55.26,5.13,1.75],[55.27,4.89,-3.24],[60.51,-29.07,3.42],[60.53,-29.31,-1.57],[13.36,-1.54,-3.08],[13.35,-1.3,1.91],[18.62,-35.74,-1.42],[18.6,-35.5,3.57]]]}']}
+tpd_template = '{"PROB_TYPE":"comp", "PROB_NAME":"NONAME", "ETA": "0.4", "DOF_PN": "3", "VOL_FRAC": "0.3", "FILT_RAD": "1.5", "ELEM_K": "H8", "NUM_ELEM_X":"10", "NUM_ELEM_Y":"10", "NUM_ELEM_Z":"10", "CHG_STOP":"0.01", "NUM_ITER":"50", "FXTR_NODE_X":"", "FXTR_NODE_Y":"", "FXTR_NODE_Z":"", "LOAD_NODE_X":"", "LOAD_VALU_X":"", "LOAD_NODE_Y":"", "LOAD_VALU_Y":"", "LOAD_NODE_Z":"", "LOAD_VALU_Z":"", "P_FAC":"1", "P_HOLD":"15", "P_INCR":"0.2", "P_CON":"1", "P_MAX":"3", "Q_FAC":"1", "Q_HOLD":"15", "Q_INCR":"0.05", "Q_CON":"1", "Q_MAX":"5"}';
 
 # TODO: fix hard coding
 # path to topy's optimization code
 rel_topy_path = './scripts/optimise.py'
 # rel_topy_path = '~/Dropbox/Projects/ProjectForte/libs/topy-master/scripts/optimise.py'
+
+offline = True
 
 #
 #   for logging timestamp
@@ -71,7 +69,9 @@ def test():
     # print on_left_side([4, 3], [65.0, 1.0], [65.0, 1.0])
 
     t0 = timestamp()
-    proc_post_data(test_data_03)
+    test_data = open('chair_01.forte', 'r')
+    # print test_data
+    proc_post_data({'forte': [test_data.read()]})
     timestamp(t=t0, msg='total time')
     return
 
@@ -142,8 +142,9 @@ def safe_retrieve_all(buffer, key, alt):
 #
 #   processing incoming data
 #
-def proc_post_data(post_data):
-    subprocess.call('rm forte_*', shell=True)
+def proc_post_data(post_data, res=192, amnt=0.15, dir=None):
+    if offline == False:
+        subprocess.call('rm forte_*', shell=True)
 
     if 'forte' not in post_data:
         # self.wfile.write('no design information')
@@ -161,9 +162,9 @@ def proc_post_data(post_data):
     #
     # read other params
     #
-    query = int(safe_retrieve_one(post_data, 'query', ANALYSIS))
-    resolution = int(safe_retrieve_one(post_data, 'resolution', 64))
-    material = float(safe_retrieve_one(post_data, 'material', 0.3))
+    query = int(safe_retrieve_one(post_data, 'query', OPTIMIZATION))
+    resolution = int(safe_retrieve_one(post_data, 'resolution', res))
+    material = float(safe_retrieve_one(post_data, 'material', amnt))
     originality = float(safe_retrieve_one(post_data, 'originality', 1.0))
     verbose = int(safe_retrieve_one(post_data, 'verbose', 0))
 
@@ -186,12 +187,12 @@ def proc_post_data(post_data):
     # compute dimensions of the voxel grid
     dim_voxel = None
     if dimension == 2:
-        dim_voxel = min((vmax[0] - vmin[0]) / resolution, (vmax[1] - vmin[1]) / resolution)
+        dim_voxel = max((vmax[0] - vmin[0]) / resolution, (vmax[1] - vmin[1]) / resolution)
     elif dimension == 3:
         print '3d is too slow currently unsupported'
         return
 
-    dim_voxel = max(dim_voxel, 2)   # avoid very tiny voxels
+    dim_voxel = max(dim_voxel, 0.1)   # avoid very tiny voxels
 
     #
     #
@@ -379,7 +380,7 @@ def proc_post_data(post_data):
 
     timestamp(msg='compute boundaries')
 
-    show_debug = True
+    show_debug = False
 
     # DEBUG: print out the design and specs
     if show_debug:
@@ -430,7 +431,10 @@ def proc_post_data(post_data):
     load_values_y_str = []
     load_nodes = [node_nums_3d(nelx, nely, 1, x[0]+1, x[1]+1, 1) for x in load_points]
     for i in xrange(0, len(load_points)):
-        node_str_array = [str(x) for x in load_nodes[i]]
+        # node_str_array = [str(x) for x in load_nodes[i]]
+
+        # instead of 8 nodes, only keep 1
+        node_str_array = [str(load_nodes[i][0])]
 
         str_load_points += ';'.join(node_str_array)
         if i < len(load_points) - 1:
@@ -449,6 +453,7 @@ def proc_post_data(post_data):
 
     tpd = json.loads(tpd_template)
     tpd['PROB_NAME'] = 'forte_' + str(long(time.time()))
+    tpd['VOL_FRAC'] = material
     tpd['NUM_ELEM_X'] = nelx
     tpd['NUM_ELEM_Y'] = nely
     tpd['NUM_ELEM_Z'] = 1
@@ -461,6 +466,10 @@ def proc_post_data(post_data):
     tpd['LOAD_VALU_Y'] = ';'.join(load_values_y_str)
     tpd['ACTV_ELEM'] = ';'.join([str(elm_num_3d(nelx, nely, 1, x[0]+1, x[1]+1, 1)) for x in actv_elms])
     tpd['PASV_ELEM'] = ';'.join([str(elm_num_3d(nelx, nely, 1, x[0]+1, x[1]+1, 1)) for x in pasv_elms])
+
+    if query == ANALYSIS:
+        tpd['CHG_STOP'] = ''
+        tpd['NUM_ITER'] = 1
 
     str_tpd = '[ToPy Problem Definition File v2007]\n'
     for var in tpd:
@@ -477,7 +486,8 @@ def proc_post_data(post_data):
     out_path = ' '  # timestamped temp output path
 
     # subprocess.call('python ' + rel_topy_path + ' ' + tpd_path + ' ' + str(query), shell=True)
-    main([rel_topy_path, tpd_path, str(query)]);
+    probname = main([rel_topy_path, tpd_path, str(query)]);
+    subprocess.call('mv ' + probname + '* ' + tpd_path + ' ' + dir, shell=True)
 
     timestamp(msg='topy')
 
@@ -545,11 +555,11 @@ def run(server_class=HTTPServer, handler_class=S, port=80):
 if __name__ == "__main__":
     from sys import argv
 
-    if len(argv) != 3:
+    if len(argv) != 2:
         print 'test mode'
         test()
     	# print 'usage: ./topy_server.py [port num] [abs path to topy optimize.py]'
     	exit()
 
-    rel_topy_path = argv[2]
     run(port=int(argv[1]))
+    offline = False
